Add: Postgres database support [GS-194]. Add: Makes for Postgres: ``` make generate_postgres_dev_sql make create_postgres_dev_tables make generate_cf_postgres make deploy_postgres ``` Add: MySQL database support [GS-249]. Add: Makes for MySQL: ``` make generate_mysql_dev_sql make create_mysql_dev_tables make generate_cf_mysql make deploy_mysql ``` Add: "make create-supad" to create the initial super admin user (supad) for local development environment [GS-125]. Change: - The "mongo" and "postgres" folders were renamed to more appropriate names, as they are now used for several databases: "mongo" is now "local_db", including local Docker containers for MongoDB, DynamoDB, Postgres, and MySQL. "postgres" is now "sql_db" because it works for Postgres and MySQL. [GS-249]: ``` /postgres -> /sql_db /postgres/generate_postgres_cf -> /sql_db/generate_sql_db_cf /postgres/generate_postgres_cf/run-postgres-deploy.sh -> /sql_db/generate_sql_db_cf/run_sql_db_deploy.sh
/mongo/run_mongo_docker.sh -> /local_db/run_local_db_docker.sh
/mongo/mongodb_stack_for_test.yml -> /local_db/local_db_stack.yml

make mongo_docker -> make local-db-up
make mongo_docker_down -> make local-db-down
make mongo_logs -> make local-db-logs
set_chalice_cnf.sh mongo_docker -> set_chalice_cnf.sh local_db_docker
```
