+++ scripts/sql_db/generate_sql_db_cf/generate_sql_db_cf.sh
@@ -1,10 +1,14 @@
 #!/bin/sh
-# generate_postgres_cf.sh
+# generate_sql_db_cf.sh
 # 2025-11-30 | CR
 # Generates the PostgreSQL tables SQL files.
+#
 # Usage:
-# sh scripts/postgres/generate_postgres_cf/generate_postgres_cf.sh
-# ACTION=generate STAGE=dev make generate_cf_postgres
+# sh scripts/sql_db/generate_sql_db_cf/generate_sql_db_cf.sh
+#    or
+# DB_TYPE=mysql ACTION=generate STAGE=dev make generate_cf_sql_db
+#    or
+# DB_TYPE=postgres ACTION=generate STAGE=dev make generate_cf_sql_db
 
 yes_or_no() {
   read choice
@@ -37,11 +41,19 @@ fi
 if [ "${STAGE}" = "" ]; then
     STAGE="$2"
 fi
+if [ "$3" != "" ]; then
+    DB_TYPE="$3"
+fi
 
 if [ "${ACTION}" = "" ]; then
     # "generate": generate the PostgreSQL SQL files.
     # "create_tables": create the tables in the local Docker PostgreSQL instance.
     ACTION="create_tables"
+fi
+
+if [ "${DB_TYPE}" = "" ]; then
+    echo "ERROR: DB_TYPE not set. Options: postgres, mysql"
+    exit_abort
 fi
 
 GT_ACTION=${ACTION}
@@ -52,6 +64,7 @@ echo "Generating PostgreSQL tables"
 echo ""
 echo "GT_ACTION: ${GT_ACTION}"
 echo "GT_STAGE: ${GT_STAGE}"
+echo "DB_TYPE: ${DB_TYPE}"
 echo ""
 
 set -o allexport; . .env ; set +o allexport ;
@@ -60,16 +73,22 @@ REPO_BASEDIR="`pwd`"
 cd "`dirname "$0"`" ;
 SCRIPTS_DIR="`pwd`" ;
 
-WORKING_DIR="/tmp/generate_postgres_cf"
+WORKING_DIR="/tmp/generate_sql_db_cf"
 mkdir -p "${WORKING_DIR}"
 
 docker_dependencies
 
+echo ""
+echo "Creating virtual environment"
+echo ""
 python -m venv venv
 . venv/bin/activate
 
+echo ""
+echo "Installing dependencies"
+echo ""
 if [ ! -f requirements.txt ]; then
-    pip install psycopg2-binary pyyaml
+    pip install pyyaml psycopg2-binary mysql-connector-python
     pip freeze > requirements.txt
 else
     pip install -r requirements.txt
@@ -81,6 +100,10 @@ if [ "${BASE_CONFIG_PATH}" = "" ]; then
     exit 1
 fi
 BASE_CONFIG_PATH="${REPO_BASEDIR}/${BASE_CONFIG_PATH}"
+
+echo ""
+echo "BASE_CONFIG_PATH: ${BASE_CONFIG_PATH}"
+echo ""
 
 if [ "${APP_NAME}" = "" ]; then
     echo 'ERROR: 'APP_NAME' environment variable not set'
@@ -100,7 +123,7 @@ fi
 
 APP_NAME_LOWERCASE=$(echo ${APP_NAME} | tr '[:upper:]' '[:lower:]')
 CF_TEMPLATE_NAME="cf-template-postgres.yml"
-FINAL_TARGET_DIR="${REPO_BASEDIR}/scripts/postgres"
+FINAL_TARGET_DIR="${REPO_BASEDIR}/scripts/sql_db"
 
 STAGE_UPPERCASE=$(echo ${GT_STAGE} | tr '[:upper:]' '[:lower:]')
 POSTGRES_PREFIX=$(eval echo \$POSTGRES_PREFIX_${STAGE_UPPERCASE})
@@ -112,7 +135,7 @@ ERROR="0"
 DONE="0"
 if [ "${GT_ACTION}" = "generate" ]; then
     # "generate": generate the CloudFormation yaml file.
-    if ! python -m generate_postgres_cf "${BASE_CONFIG_PATH}" "${WORKING_DIR}/${CF_TEMPLATE_NAME}" "" "0"
+    if ! python -m generate_sql_db_cf "${BASE_CONFIG_PATH}" "${WORKING_DIR}/${CF_TEMPLATE_NAME}" "" "0" "${DB_TYPE}"
     then
         # We need to postpone the error reporting to remove __pycache__ and venv every time the python code is executed.
         ERROR="1"
@@ -124,11 +147,11 @@ fi
 if [ "${GT_ACTION}" = "create_tables" ]; then
     # "create_tables": create the tables in the local Docker postgres instance.
     
-    # Verify if postgres local container is running to avoid cycling "make mongo_docker" call
+    # Verify if postgres local container is running to avoid cycling "make local-db-up" call
     if ! ${DOCKER_CMD} ps | grep postgres-local -q
     then
         cd "${REPO_BASEDIR}"
-        if ! make mongo_docker
+        if ! make local-db-up
         then
             echo ""
             echo "ERROR: Failed to start the local Docker databases container"
@@ -137,7 +160,7 @@ if [ "${GT_ACTION}" = "create_tables" ]; then
     fi
 
     cd "${SCRIPTS_DIR}"
-    if ! python -m generate_postgres_cf "${BASE_CONFIG_PATH}" "${WORKING_DIR}/${CF_TEMPLATE_NAME}" "${POSTGRES_PREFIX}" "1"
+    if ! python -m generate_sql_db_cf "${BASE_CONFIG_PATH}" "${WORKING_DIR}/${CF_TEMPLATE_NAME}" "${POSTGRES_PREFIX}" "1" "${DB_TYPE}"
     then
         # We need to postpone the error reporting to remove __pycache__ and venv every time the python code is executed.
         ERROR="1"
@@ -154,10 +177,10 @@ rm -rf venv
 if [ "${GT_ACTION}" = "generate" ]; then
     if [ "${ERROR}" = "1" ]; then
         echo ""
-        echo "ERROR: Failed to generate SAM template file"
+        echo "ERROR: Failed to generate CloudFormation/SQL template file"
     else
         echo ""
-        echo "Generated SAM template file in:"
+        echo "Generated CloudFormation/SQL template file in:"
         echo "${WORKING_DIR}/${CF_TEMPLATE_NAME}"
         echo ""
         echo "Do you want to edit it (y/n)?"
@@ -218,10 +241,10 @@ fi
 if [ "${GT_ACTION}" = "create_tables" ]; then
     if [ "${ERROR}" = "1" ]; then
         echo ""
-        echo "ERROR: Failed to generate the Postgres tables in the local environment"
+        echo "ERROR: Failed to generate the Database tables in the local environment (${DB_TYPE})"
     else
         echo ""
-        echo "Generated Postgres tables in the local environment:"
+        echo "Generated Database tables in the local environment (${DB_TYPE})"
         echo ""
         ENDPOINT_URL="${POSTGRES_URI}"
         if [ "${ENDPOINT_URL}" = "" ]; then
